requirements.txt
python-dotenv
black
ruff
pytest
pg8000
asyncio
asyncpg